fix(ci): keep OpenAPI specs lintable
